Update kernel.cu
Added proper global kernel support, significantly trimmed down abstraction bloat, reduced complexity and ported to raw pointers for the values and hash tables as well.  Added overrides standard for unified memory contexts that allow for the new keyword, as well as the delete keyword. Refactored generally to make the map operate as a pointer and to be dealt with via CudaDeviceReset properly.  Acts as a minimal overhead fixed-size hashmap for the purposes of accessing and modifying data without having to frequently linearize and repack data between global kernel calls and ends.